+++ ReleaseNotes.md
@@ -2,7 +2,9 @@
 **Version 1.0.6 (compatible with SurrealDB version 2.0._):**
 - versionstamp is now can be used in SINCE statement for SHOW
 - by default, if no SINCE were specified Show statement generates SINCE 1
+- add DEFINE CONFIG statement, only Database object can use it
 - add examples and tests for Show
+- add examples and tests for DEFINE CONFIG
 
 **Version 1.0.5 (compatible with SurrealDB version 2.0.4):**
 - minor examples fix